Move the main plugin to the application folder
